abilities added

+++ Assets/Source/Units/Enemies/IMeleeAtacker.cs
@@ -1,4 +1,4 @@
 public interface IMeleeAtacker
 {
-    public abstract void Atack(ushort damage);
+    public abstract void Atack(int damage);
 }
