+++ Assets/Source/Units/Recruit/Specializations/Warrior/Warrior.cs
@@ -1,14 +1,9 @@
-using System;
-using System.Collections;
-using UnityEngine;
-using UnityEngine.Events;
-
 public class Warrior : Recruit
 {
     public float _damageBonus = 1.5f;
 
-    public override void Atack(ushort damage)
+    public override void Atack(int damage)
     {
-        base.Atack((ushort)(damage * _damageBonus));
+        base.Atack((int)(damage * _damageBonus));
     }
 }
